+++ novelservice/src/main/java/com/iucyh/novelservice/common/exception/GlobalExceptionHandler.java
@@ -2,12 +2,12 @@ package com.iucyh.novelservice.common.exception;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.iucyh.novelservice.common.dto.apiresponse.ApiResponseMapper;
-import com.iucyh.novelservice.common.dto.apiresponse.information.FailInformation;
+import com.iucyh.novelservice.common.response.api.ApiResponseMapper;
+import com.iucyh.novelservice.common.response.api.information.FailInformation;
 import com.iucyh.novelservice.common.exception.errorcode.CommonErrorCode;
 import com.iucyh.novelservice.common.exception.errorcode.ErrorCode;
 import com.iucyh.novelservice.common.util.IpUtil;
-import com.iucyh.novelservice.common.dto.apiresponse.FailResponse;
+import com.iucyh.novelservice.common.response.api.FailResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
