+++ novelservice/src/main/java/com/iucyh/novelservice/common/response/api/information/FailInformation.java
@@ -1,4 +1,4 @@
-package com.iucyh.novelservice.common.dto.apiresponse.information;
+package com.iucyh.novelservice.common.response.api.information;
 
 import com.iucyh.novelservice.common.exception.errorcode.ErrorCode;
